Starting the long process of converting and open-sourcing Twelve Sands.  :)
I'm excited to be moving ahead with this, after so many years.  It's time
to get the game out there to give people a chance to play with it on
their own terms.

With that said, Twelve Sands was written with an older version of PHP,
and needs some attention before it goes live.  It's tied relatively closely
to the server it was built on, so I'm going to decouple it and get it
running on a local MAMP server before pushing it live.