More main directory include files, formatted.
A lot of the custom artifacts, zones, and other pieces of the game are
either data driven (database) or handled using these custom pieces of code
that are included only when necessary.  For example, the books and pages
in the game are covered in _read.php.  The artifact use code will include
this file for readable types.