Immediate action page (action/redirect), and Bank of Nobility.
The action page is responsible for activities that we don't want to
handle as part of an existing activity.  For example, some
modifications to the user object may be better off done on their own,
followed by a full reload.  In some cases, it's just easier to
tackle it on its own, then send a header notification to get to the
new area.