Asynchronous duel code.
Of all the things that Twelve Sands was able to do, this is one of
the things that I'm most proud of.  It's not a gigantic technical
achievement (although it is kind of cool), but the fact that you
could have real-time duelling with full state management against
another player online at the time was satisfying.  Some of my
best memories of actually _playing_ the game were the real-time
duels.

Later changes to the combat engine made some of this redundant,
and it might not be in a fully functional state now.  However,
if it's not, I'm certainly going to get this up and working again.