Enchanting interface.
This was another big addition to the engine that I was really proud
of.  Enchanting was a concession to database storage, and required
a considerable amount of rewriting in the code to support what was
essentially a second index (artifact identifier and enchantment
identifier).  I think it was worth it though.  This was never tested
against huge amounts of data or players, but I'm fairly confident
that the code was resilient enough to support a larger player base.
Plus, it's cool to be able to sink artifacts for gems, and to add
premade enchants to any artifact of a particular type.  Another
one of my favourite parts of developing the game.  :)