Achievements support.
One of the most interesting additions to Twelve Sands, from a
player perspective.  When Twelve Sands was unable to provide
enough content for players (i.e., when players were so interested
in the game that they blasted through everything I could offer
faster than I could generate new content), achievements gave them
something meaningful to work toward.

I really like this system, in its efficiency and in how easy it
was to add new achievements.  A few lines of code and a database
entry, and the achievement was live.