State files.
These are necessary for the cases in which we want to do some
significant processing before the character header is launched.  By
moving logic out of the in-page rendering and doing everything before
we start to push the html code, we know the character object is up to
date with the latest action.